backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
PyJWT
pydantic==2.5.1
motor==3.3.2
python-dotenv==1.0.0
httpx==0.25.0
requests==2.31.0
